+++ python-package/HISTORY.md
@@ -6,6 +6,10 @@ Unreleased
 
 -- see the [milestone tracking at github](https://github.com/k4cg/nichtparasoup/milestone/2).
 
+* Breaking Changes
+  * Imagecrawler for Instagram is no more part of standard distribution.
+    It might come back as a plugin one day.  
+    This crawler was just too brittle because of Instagram's WebApplicationFirewalls and other bot protections.
 * Changes
   * API supports HTTP method "GET" only. Did support all HTTP methods in the past. 
 * Added
@@ -14,7 +18,7 @@ Unreleased
   * API `/get` no longer responds false "404 EXHAUSTED" HTTP Status code.
   * `nichtparasoup.server.get_image()` no longer responds false `None`.
 * Removed
-  * Crawler `nichtparasoup.imagecrawlers.instagram.InstagramProfile` no longer has a new optional config `profile_id`.  
+  * Crawler `nichtparasoup.imagecrawlers.instagram` was removed from shipped imagecrawlers.
 
 ## 3.0.0a2